move infomenu to lsb dir
